Corrigida issue #53
